Separated the filtering options and removed filtering for Features and Categories
